Figured out problem with wayfinding + Added some rooms and paths
Problem with wayfinding was that anytime a path splits off from another path, the path must end where the other splits off and then another must start at the end of it. In english: Paths must be added at anchor points, never in the middle of a path. Otherwise the path will not be seen by the wayfinding plugin.