Wayfinding: Added Scroll Bar + Refresh Screen button
Added scrolling functionality with both a scrollbar and scrolling the rooms by dragging. The scroll bar is kinda finiky though, If you scroll the scroll bar AFTER already scrolling the room numbers the location of the scroll bar gets all out of wack. Also added a refresh screen button to the bottom left corner of the wayfinding page so users can refresh the page if they think there is a problem.