Added group name edit + updated plans
Added ability to edit a groups name, along with other features you would expect like exiting the edit mode when escape is pressed. Also updated plans.txt with what is left to do on the project, once those are complete i think it is safe to say the app is complete.