Fixed + added lots of stuff
Relaized i could just include php for any ad that isnt just an iframe or an image. Removed all refrences to sub content, which i created in order to have overlays, this has moved on to only be used in PHP scripts. Also added some ramfest stuff for campus activites.