Converted PDF's to JPGs + Added <video>'s + reordering of Ads in group on status page
Converted some PDF's to JPG's, why the hell would i want to display a PDF?? Theni allowed for local videos to be played by adding a video tag which is also handled in the JS to ensure it only plays when it is supposed to. Lastly, I added a feature which allows you to reorder ads in a a particular group by dragging them on the status page and clicking the save button. The save button sends a  asyncronous request to the server telling it to reorder the ads. Cool stuff!