Added slide fade + fixed revslider iframes (kinda)
Added a cross fade to slides. This involved making the opacity of the .Ad_Content class zero, and its postion to be absolute. then i used the jQuery fadeTo method to simply fade the previous ad into the next ad. Also defined iframes to be the size of the window. This needs work, should be checked right before the ad plays to ensure it is the correct size.