Reduced Weather API calls + Added dependencies + add Variables.csv
Redueced the number of weather API calls by moving the value into a file with an expiration time. Now, instead of EVERY screen / "instance" performing an API call every 15 minutes we check if the temperature is up to date if it hasnt been updated within the last 15 minutes. This should pull us down to a meager ~2,600 weather API calls / month instead of over 5,000, which requres a paid monthly plan.

the Dependencies folder allows me to throw things like phpflickr into a more organized place.

The Variables.csv is where i plan to house variables which make sence to only be updates every so often. For example, temperature. This file is actually meant to reduce the number of API calls to any given API.