Added random content playback globably
Added a GLOBAL modifier for content to be played back in a 'random' way. Note: groups are still played in order, but the group content is random. So if you have a group with {1,2,3} and a group with {4,5,6,} and you randomize them both you would have something like {3,1,2} and {5,6,4}. what im getting at here is group content will always stick together and wont be scattered through other junk.